collector: Improve failure and clean-up
The class destructor __del__ is only called when all references to the
object cease to exist, which in this is case if after the program exits.
It is better to explicitly call it instead, before the program exits,
and only when we actually need to perform a clean-up.

This commit also reorganizes other things a bit, initializing daemon,
unit_was_runing and defining the eos-upload-metrics path in the
contructor, and cleans-up some other details.

https://phabricator.endlessm.com/T30684