collector: unmask the metrics daemon to prevent start over D-Bus
The metrics daemon also needs to be masked to gurantee it won't be
started until the collect/upload is done. It simply removes the
link to achieve the `unmask`. The service should be activated via
D-Bus at some point and we don't really have to start it right after
collect/upload done.

https://phabricator.endlessm.com/T30975