daemon: Add internal API to enqueue aggregate events
emer_daemon_record_aggregate_event() was previously called directly by
the handler for the RecordAggregateEvent D-Bus method. This D-Bus method
has been made a no-op since the format of aggregate events has changed
in the new wire protocol.

Adjust emer_daemon_record_aggregate_event() to add a "complete"
agggregate event to the buffer of events awaiting submission to the
server, and re-enable the test code which checks that these are
correctly sent on the wire.

There are no non-test users of this API yet. It will be used in future
by code that runs at the end of each day and month to enqueue all
aggregate events for that day/month from a separate buffer of
"incomplete" events.

https://phabricator.endlessm.com/T32203