daemon: Introduce aggregate timers
This is split as follows:

 * On startup, EmerDaemon sets up a boottime-based clock
   ticking a callback at midnight, and creates the
   EmerAggregateTally object.

 * emer-main.c receives and handle the D-Bus call, create
   the D-Bus timer object, and fetch the client's unix user
   id. It also notifies EmerDaemon about the new aggregate
   timer.

 * When a timer should be created, EmerDaemon creates a new
   EmerAggregateTimerImpl object for this timer, adds it to
   its internal hash table, and monitors the D-Bus timer for
   the "StopTimer" signal.

 * When a timer is stopped, it is stored both at the daily
   and monthly tallies.

Every midnight, the daily tally is buffered for submission.
If the month changed, the monthly tally is also buffered for
submission. After the tallies are buffered, all ongoing timers
are split, and their counter is reset.

This commit also lists com.endlessm.Metrics.AggregateTimer to
the D-Bus allowlist.

Incoming tests next.