tally: Use sqlite3 for storage
I have long thought it would be nice to use a real database for
this daemon's on-disk cache. Here is a draft of how it would look
to switch EmerAggregateTally to use sqlite.

The change in daemon.c:

+    emer_aggregate_tally_new (priv->persistent_cache_directory ?: g_get_user_cache_dir ());

is needed because the test passes a MockPersistentCache rather
than passing a persistent_cache_directory to the EmerDaemon, so
persistent_cache_directory is empty. This should be cleaned up
somehow.