configure: Don't require --with-default-metrics-server
2fa2c669ed5e313491f23579ea8a9af0791a0d48 (in
https://github.com/endlessm/eos-event-recorder-daemon/pull/249)
introduced the ability to configure the metrics server URL at build
time, and removed the default being to submit to the Endless server, to
ensure that development builds and builds for other distros don't
accidentally send metrics to Endless.

At the time, --with-default-metrics-server was made mandatory, with the
argument that people don't look at warnings from configure scripts.
While true, I think the other downsides outweigh this. It is desirable
that a project uses a standard build system and can be successfully
built and tested out of the box. Every time I clone this repo, I am
annoyed that I have to set this flag, and have to think about what value
to use during development. We had to hardcode a dummy value in the build
system for 'make distcheck' to pass, and another value in our Jenkins
configuration to allow it to configure the project in the first place in
order to run 'make distcheck'.

Now that I find myself wanting to switch this project to Meson, the
Jenkins stuff is an active impediment, because Meson doesn't accept the
same configuration arguments, so I would need to update the Jenkins
config in lockstep and find some way to maintain build support for older
OS branches.

Instead, let's define an obviously-dummy value as the default, the same
one as used by 'make distcheck', and emit a warning. 'azafea.example'
will never resolve and is obviously a placeholder, so you'll notice soon
enough in the journal that no metrics are being submitted (and they'll
be buffered on disk).

https://phabricator.endlessm.com/T32969