tally: Store UUIDs as 16-byte BLOBs
Previously, event IDs were stored in the GVariant text serialization of
an array of bytes:

    [byte 0x49, 0xd0, 0x45, 0x1a, 0xf7, 0x06, 0x4f, 0x50, 0x81, 0xd2, 0x70, 0xcc, 0x0e, 0xc9, 0x23, 0xa4]

Instead, store those 16 bytes in a BLOB.

The other option would have been to store the string representation,
e.g. 49d0451a-f706-4f50-81d2-70cc0ec923a4. This is more legible when
using the sqlite3 CLI for example, which otherwise just emits the bytes
directly to the terminal. But we already have that problem for the
serialized GVariants stored in other fields of the table. Storing the
UUID bytes is the densest possible representation and allows a very
simple CHECK constraint in the schema.

As part of this change I opted to use uuid_t to represent event IDs
throughout the tally code, converting on the way in and out. It seems
cleaner and the diffstat agrees.

https://phabricator.endlessm.com/T33140