tests: Simplify aggregate-tally test
Constructing and tearing down an AggregateEvent struct every time we
insert an event into the tally ended up as boilerplate that (to my eye
at least) obscured what was going on. Inline the four variables into
each function.

I also took the opportunity to hoist some repeatedly-defined variables
to the top level, and initialize things where they are defined, for
similar reasons: I think this makes the important part of the function
more legible.

I kept the struct because it will be more useful, in expanded form, when
checking the data that comes back out of the tally!