tally: Give database a generic filename
My long-term hope is that we remove the bespoke on-disk ring buffer from
this project in favour of storing all cached events in the same SQLite
database.

To this end, give the database file a generic name.  This also avoids
the need to handle a schema migration due to the schema changes over the
past dozen commits.

https://phabricator.endlessm.com/T33140