main: Update StartAggregateTimer signature
The method now includes a caller-specified Unix user ID, like the older
RecordSingularEvent() method and its friends. This allows us to remove
the aggregate key, since it was only used to disambiguate by user ID in
the case where the events are actually being reported by root.

Until the tally class is updated to not need an aggregate key, pass the
payload in its place.

I considered whether to make the daemon continue to check the caller's
real UID, and (e.g.) only allow root to specify a UID that is not their
own. But I don't think this is necessary. The UID is only used for
disambiguation, not for any security purpose, and the worst that happens
is that someone spoofs app usage for another local user, or even a
non-existent user. But this is already the case for singular events, and
this is just telemetry at the end of the day.

https://phabricator.endlessm.com/T33252