tally: Don't store aggregate_key
This was previously only distinct from the payload in order to
disambiguate events by Unix UID without them being recorded by that UID
and also without sending the UID to the server. Now that the D-Bus API
allows the caller to supply a UID, we don't need to store a separate
aggregate key.

https://phabricator.endlessm.com/T33252