tests: Test calling StartAggregateTimer over D-Bus
Most of the EmerDaemon class knows nothing about D-Bus, with the binding
of C functions to corresponding D-Bus API happening in emer-main.c. This
one is an exception: among other things it accepts a bus name as a
parameter, and EmerDaemon watches for that name falling off the bus.
This means it cannot easily be tested in test-daemon.c like other
methods.

Exercise it in the existing opt-out integration test, which talks to the
daemon over D-Bus. Ideally it could be refactored so that most of the C
logic could be tested from C, but this is better than nothing. And one
of the tests is that it returns an error when metrics are disabled, so
it seems reasonable to put that here.

https://phabricator.endlessm.com/T32765