daemon: Don't double-wrap payloads in variant
The D-Bus method RecordSingularEvent() accepts type 'v' for its payload
argument, and passes its arguments directly to
emer_daemon_record_singular_event(). So the GVariant *payload argument
is always of type 'v'.

emer_daemon_record_aggregate_event() was previously exposed as
RecordAggregateEvent(); these days it is called internally, but with a
GVariant that was originally received over D-Bus as type 'v'. So again,
the parameter is always of type 'v'.

But when g_variant_new() encounters a 'v' format character, it always
wraps the GVariant * argument it consumes in another layer of type 'v'.

This means that, when the application passes a payload (in GVariant text
notation) to the daemon as:

    <'my beautiful payload'>

It is sent to the metrics server as:

    <<'my beautiful payload'>>

This has no practical impact since Azafea has a loop to unwrap as many
layers of variant as necessary to feast on the goo inside, but it is
unsightly and (in my opinion) confusing.

Fix this by asserting that the GVariant passed to
emer_daemon_record_[...]_event() is of type 'v', and then not boxing it
again. The tests need a small corresponding adjustment.