daemon: Don't critical on aggregates with NULL payload
emer_daemon_record_aggregate_event() was previously exposed as a D-Bus
method, RecordAggregateEvent(), which was superseded by the aggregate
timer mechanism.

Because the GVariant 'maybe' type is not available on D-Bus, the D-Bus
method and hence this function simulates it with a boolean plus payload;
the payload is ignored if the boolean is false, but the payload itself
was always non-NULL.

However, this function is now called with a GVariant loaded from the
tally, which is NULL if there is no payload. The conversion from
boolean + variant to nullable variant occurs when StartAggregateTimer()
is called. As a result, get_nullable_payload() would log a critical when
an event with NULL payload is enqueued. (As far as I can tell this is to
guarantee that a floating GVariant passed as payload to
emer_daemon_record_*_event() is sunk, which in turn is only for the
benefit of the test suite.)

Rename the C function to make clear that its meaning has changed, remove
its has_payload argument, and rely on C-level nullability of the
payload.

https://phabricator.endlessm.com/T33299