tests: Use r'' string for regular expression
The intent of this code is to match a literal '.', not the any-character
wildcard, in the name of the D-Bus error. But without prefixing the
string with r'' to mark it as a raw string, the \ is interpreted as a
string escape.

Since '\.' is not a valid escape sequence, this works, but flake8 warns:

    tests/test-opt-out-integration.py:96:61: W605 invalid escape sequence '\.'
    [...]

Mark these as raw strings.