Store running timers on shutdown
On system shutdown, systemd sends a signal to the daemon to shut down.
If there any timers running at this time, previously they would not be
stored and the data would be lost.

In some cases, it may be that the process which owns the timer may stop
its timers or be killed before the daemon shuts down, with enough time
before the daemon is signalled for it to react and store the timers. But
we can guarantee that the data will be saved by explicitly stopping all
timers on shutdown.

https://phabricator.endlessm.com/T32730