daemon: Correctly remove timers when client disconnects
priv->aggregate_timers is a hash table with char * keys and
EmerAggregateTimerImpl * values. The key is a property of the
EmerAggregateTimerImpl.

Previously, on_timer_stopped_cb(), which implements the StopTimer() D-Bus method, correctly removed the running timer from this hash table.

But sender_name_vanished_cb(), which handles the case where the client
that started one or more timers disconnects from the bus before calling
StopTimer(), did not: it incorrectly passed the EmerAggregateTimerImpl *
as the key to remove. As a result, a warning was logged in this case for
each running timer, and (I think) the timer would continue conceptually
"ticking" and its balance would be added to the tally every midnight
until the daemon/computer happened to restart.

Factor out a remove_timer() function, which correctly removes the timer
from the hash table, and use it in both places.

https://phabricator.endlessm.com/T32730