test-persistent-cache: Don't use hardcoded test directory
I noticed when telling Meson to run tests 1,000 times in a loop to try
to find races that this test uses a hardcoded directory in /tmp, which
among many other negative properties means that it can't be run twice at
once.

Happily GLib has a built-in way to create a fresh directory for every
test case and delete it after the test is run. Sadly, using it involved
a massive diff.

https://phabricator.endlessm.com/T32730