Simplify handling overlapping timers
Previously, if a client made a second identical call to
StartAggregateTimer() while the first was still alive, the daemon would
return the existing timer, incrementing a reference count. Each call to
StopTimer() would decrement that refcount, and only stop the timer once
the refcount reached zero. This avoided double-counting the overlapping
time, at the cost of some complexity in the daemon to maintain the
reference count.

On balance, I don't think the upside is worth the complexity. In the
known case where this situation unavoidably occurs, the overlap is only
for as long as it takes for a pair of D-Bus messages to be exchanged
between gnome-shell and the metrics daemon: negligible.

Remove the refcounting, and return a fresh timer each time.

https://phabricator.endlessm.com/T32730