Introduce a clock abstraction
Currently, test-daemon works by setting the upload frequency to "every 2
seconds", and waiting for that time to elapse before the daemon sends
requests to the mock server. As a result, this test takes around 45
seconds to run. (I guess we can infer that it waits for 22 requests to
be sent.)

I have tried reducing the timeout to 1 or 0, but this is racy: the
daemon will happily send a request even if there are no events to send.
(Arguably a bug but one for another day.)

In preparation for fixing this more robustly, introduce an EmerClock
interface. The production daemon will use one implementation of this,
EmerRealClock, and the test suite will provide a mock implementation,
MockClock.

Subsequent patches will add methods to this interface and adjust
emer-daemon.c to use them rather than the underlying GLib functions
directly.

https://phabricator.endlessm.com/T33952