daemon: Mock g_timeout_add_seconds in test suite
Prevously, test-daemon.c would take over half a minute to run. This is
because it set the submission interval to 2 seconds, and simply waited
for that time to elapse each time it wanted to test the daemon sending a
request to the server.

Add a function to the EmerClock interface that abstracts
g_timeout_source_new_seconds(), and use it in EmerDaemon. The real
implementation is trivial; the mock implementation, rather less so.

With this change, the test takes about 1.5s on my machine, down from
around 42s.

https://phabricator.endlessm.com/T33952