daemon: Emit a signal on transient upload failure
Previously, EmerDaemon would call g_warning() and the test suite would
listen out for that.  GLib's esteemed developers discourage libraries
from logging to stderr during normal operation, and tests testing this.
See https://gnome.pages.gitlab.gnome.org/gtk/glib/logging.html#testing-for-messages

Instead, have EmerDaemon emit a new ::transient-upload-error signal;
catch this at the top level of the application and log a warning as
before. Adjust the test suite to await the emission of this signal
rather than asserting about log output.

Previously, the message was logged unconditionally before deciding
whether to schedule a retry, or fail the operation. The new signal is
emitted only when the daemon has chosen to schedule a retry. (This will
help with another improvement to this test code shortly, which needs to
know when a retry has been scheduled.) This means that in the case where
an HTTP request is cancelled locally while in flight, and the server
subsequently returns an error, no warning will be logged when previously
it would have been. A minor and I believe unimportant behaviour change,
but I'm noting it here for completeness.

https://phabricator.endlessm.com/T33952