test-persistent-cache: Avoid VLA
num_variants is, in practice, 303 030, so this array is 303 030 * 8 =
2 424 240 bytes, a little over 2 MiB.

When run under valgrind, this test would show:

    ==55772== Warning: client switching stacks?  SP change: 0x1ffefff2f0 --> 0x1ffedaf540
    ==55772==          to suppress, use: --max-stackframe=2424240 or greater

followed by thousands of "Invalid read/write of size 8" warnings for
indexing into this array.  This is because, by default, Valgrind assumes
a stack frame will be no larger than 2 000 000 bytes.

We could follow the advice to grow the stack size; but it's easier to
just heap-allocate this array.

https://phabricator.endlessm.com/T33959