tests: Remove launch-mock-dbus-tests.sh
test-daemon does not actually exercise any code in EmerDaemon that uses
the bus. (Some aggregate timer-related codepaths do use the bus, but
they are not tested by this test, and ideally the D-Bus stuff for timers
would be moved out to emer-main.c like it is for the rest of the
daemon.)

Verify this by specifying /dev/null as the session and system bus
sockets, and remove the wrapper script that runs a temporary bus.

Removing the wrapper script makes it easier to run the test itself under
valgrind (rather than the shell running the wrapper script) and feast on
the errors discovered.

https://phabricator.endlessm.com/T33959