aggregate-tally: Sink a passed-in floating GVariant
Valgrind detected that test_aggregate_tally_store_events leaks its local
'GVariant *payload' variable. This is because v_str() returns a strong ref.

As it happens, no floating refs are passed in, either in real code or in
test code, but it's easy enough to make it work either way and test
that.

https://phabricator.endlessm.com/T34115