tests: Run daemon subprocess under Valgrind
test-opt-out-integration.py works by spawning the full daemon as a
subprocess, after having spun up a temporary bus & arranged for some
mocks to exist on it.

As a result, the current '--setup valgrind' support valgrinds the Python
test process, not the daemon.

Place the (possibly-empty) command to wrap the daemon subprocess with in
an environment variable, so that the test can apply it when spawning the
daemon.

As noted in the comment, Valgrind's own --trace-children=yes is not a
good option because the test spawns other subprocesses, such as a D-Bus
daemon, and we don't want to pay the cost of that or get bogged down in
suppressing known leaks there etc.

https://phabricator.endlessm.com/T34115