persistent-cache: Remove code to unlink really old files
In 486994262cdd347945f0b67bc165969fc117088b
"Overhaul persistent cache API.", the persistent cache format was
changed from three separate files to a single file, and some code was
added to unlink the old files.

Valgrind reveals that this code reads uninitialized memory in the case
where emer_cache_version_provider_get_version() returns FALSE, it does
not assign anything to its second parameter, and so old_version is read
despite being uninitialized in the body of the conditional.

This commit was made in 2015 and was released with Endless OS 2.4.0. It
is a safe assumption that any machine from back then which is still
running has had plenty of opportunities to delete these files. We also
changed the persistent cache path in
88a3c477888d26a948a2202f23cd9167d5f0d8bf "daemon: Bump
CLIENT_VERSION_NUMBER to 3" and these files never existed in the new
path.

Remove this cleanup path, and initialize old_version for good measure.

https://phabricator.endlessm.com/T34115