Make the server URL configurable in the config file
This makes it easier if we want to test a different server without
needing to recompile the daemon for this.

The default server URL that is passed to meson (as
-Ddefault_metrics_server_url) is written out to the config file, but
at the same time also hardcoded as a default fallback in case clients
don't get the updated configs (this can easily happen if ostree or rpm
are managing the config files).

The default_metrics_server meson option is now renamed to
default_metrics_server_url to signify that it now accepts a slightly
different syntax: instead of azafea.example, it now takes
https://${environment}.azafea.example:443 where ${environment} variable
gets replaced by the environment variable value from the same config
file. This avoids the previous behaviour where the environment value was
magically prepended to the host address and makes it more obvious what
the actual host address that is used is going to be.