daemon: Remove server-url property
EmerPermissionsProvider now has a method to fetch the server URL. Make
this method return the build-time default value if no value is specified
in the config file, and adjust EmerDaemon to only use that method.

Extend mock-permissions-provider to return a construct-time value.

(It is a little unusual to return a freshly-allocated copy of a static
from emer_permissions_provider_get_server_url() but that is how that
function already works.)

https://phabricator.endlessm.com/T35165