Expand server URL on every upload
Previously, EmerPermissionsProvider would return the raw string value
from the config file (or a built-in default value), and EmerDaemon would
expand any ${environment} placeholder therein. EmerDaemon.server_url
would cache the result of this expansion after the first upload, so
subsequent changes to the environment (such as by changing the OSTree
server URL) would not be reflected.

Move expansion of the URL to EmerPermissionsProvider, and have
EmerDaemon call that function each time the URL is used.

The test change is necessary because of some really weird emergent
behaviour of this code. Normally Unix services read their configuration
files once at startup, when explicitly asked to re-read them
(conventionally with SIGHUP), but not at other times; and never write to
them. However, EmerPermissionsProvider is not normal:

1. It re-reads its config file every time you call a function like
   emer_permissions_provider_get_environment(), so its properties like
   EmerPermissionsProvider:enabled can change in response to you calling
   an accessor function!

2. Reading a property of EmerPermissionsProvider can cause the file to be
   written:

  a. If an error occurs when reading the keyfile, such as the keyfile not
     existing, EmerPermissionsProvider writes the defaults back to that
     file.

  b. emer_permissions_provider_get_environment() inspects the system's
     'eos' OSTree remote to see if it looks like a developer system; if
     so, and if the configured environment is 'prod', it spontaneously
     changes the environment to 'dev' (and overwrites the config file).

3. The daemon has a SetEnabled() D-Bus method which is wired up to
   emer_permissions_provider_set_enabled(), which also writes the file.

In previous incarnations of this code, calling UploadMetrics() would
cause the environment to be read prior to checking whether metrics are
enabled. This would in turn cause the config file to be written if it
didn't previously exist. And this test is relying on that behaviour to
inspect the daemon's internal state. (Interestingly the test also
claimed that the daemon never reads its config once it is started, but
in fact it relies on it doing so.)

Now, the environment is not consulted until after checking whether
metrics are enabled and determining that they are. This breaks the
weird behaviour that the test is relying on.

https://phabricator.endlessm.com/T35165