meson: Use ExternalProgram.full_path()
Meson 0.55 deprecated ExternalProgram.path() in favour of a new
function, ExternalProgram.full_path(). The two functions are identical
in behaviour, but Meson logs a "future-deprecated" warning if you use
.path(). The renaming is not explained in the deprecation message or
documentation, but appears to be solely for consistency with other
objects.

Depend on a new-enough Meson for the new .full_path() method, and use
it.

https://phabricator.endlessm.com/T35165