daemon: Disconnect signal handler on dispose
EmerDaemon listens for change notifications on
EmerPermissionsProvider:daemon-enabled. Previously, it never
disconnected this signal handler. In normal usage this is not a problem
because the EmerPermissionsProvider is not reused for a different
EmerDaemon instance. However, in the test suite it is. This manifested
itself as non-deterministic segfaults.

Add an assertion that makes the failure deterministic.  Use
g_signal_connect_object() to automatically disconnect the signal handler
when the EmerDaemon object is disposed. Depend on GLib 2.74 for the
G_CONNECT_DEFAULT constant, which is just a self-documenting way of
writing the number 0.

https://phabricator.endlessm.com/T35165